requirements.txt
# Flask Backend
flask>=3.0.0
flask-cors>=4.0.0

# DeepSeek (OpenAI-compatible SDK)
openai>=1.0.0

# PDF + DOCX Parsing
PyPDF2>=3.0.0
python-docx>=1.1.0

# OCR Fallback
requests>=2.31.0

# MongoDB
pymongo>=4.7.0
pymongo[srv]

# Environment
python-dotenv>=1.0.0

# Utilities
numpy>=1.26.0
